+++ src/plots/PlotCreateMany.tsx
@@ -6,6 +6,7 @@ import {
     FileInput,
     FileField,
     useRedirect,
+    useNotify,
 } from 'react-admin';
 import { Typography } from '@mui/material';
 import { useState } from 'react';
@@ -14,6 +15,7 @@ const PlotCreateMany = () => {
     const dataProvider = useDataProvider();
     const [errors, setErrors] = useState([]);
     const redirect = useRedirect();
+    const notify = useNotify();
     const save = async (data) => {
         try {
             const response = await dataProvider.createMany(
@@ -23,6 +25,7 @@ const PlotCreateMany = () => {
                 onSuccess(response);
             }
         } catch (error) {
+
             if (onFailure) {
                 onFailure(error);
             }
@@ -30,7 +33,16 @@ const PlotCreateMany = () => {
     };
 
     const onFailure = (error) => {
-        setErrors(error.body.detail.errors);
+        // If 400 or 500 error, show error message
+        // If the body detail is str and not list), it means that the error is
+        // not a validation error and simply provide notification
+        if (typeof error.body.detail === 'string') {
+            notify(`Error: ${error.body.detail}`, 'error');
+        }
+
+        if (error.body) {
+            setErrors(error.body.detail.errors);
+        }
     }
     const onSuccess = () => {
         redirect('list', 'plots');
@@ -42,7 +54,7 @@ const PlotCreateMany = () => {
                     <Typography variant="h6">Upload a CSV containing the plot data</Typography>
                     <Typography variant="body3">The CSV must contain the following header:</Typography>
                     <Typography variant="caption"><i>project_name,id,area_name,gradient,date,coord_z,coord_x,coord_y,slope,aspect,weather,topography,vegetation_type,lithology</i></Typography>
-                    <Typography variant="caption">Where <i>project_name</i> refers to the name of the Project that the plot will be assigned to. Duplicates are not accepted, and any errors will be shown after uploading.</Typography>
+                    <Typography variant="caption">Where <i>project_name</i> refers to the name of the Project that the plot will be assigned to. Duplicates are not allowed, and any errors will be shown after uploading.</Typography>
                     <FileInput
                         label="Plot sample data (.csv)"
                         accept=".csv"